+++ jdaviz/configs/imviz/plugins/aper_phot_simple/aper_phot_simple.py
@@ -8,7 +8,7 @@ from bqplot import pyplot as bqplt
 from glue.core.message import SubsetCreateMessage, SubsetDeleteMessage, SubsetUpdateMessage
 from ipywidgets import widget_serialization
 from regions.shapes.rectangle import RectanglePixelRegion
-from traitlets import Any, Bool, List, Unicode
+from traitlets import Any, Bool, List, Unicode, observe
 
 from jdaviz.configs.imviz.helper import layer_is_image_data
 from jdaviz.core.events import AddDataMessage, RemoveDataMessage, SnackbarMessage
@@ -22,7 +22,11 @@ __all__ = ['SimpleAperturePhotometry']
 class SimpleAperturePhotometry(TemplateMixin):
     template_file = __file__, "aper_phot_simple.vue"
     dc_items = List([]).tag(sync=True)
+    data_selected = Unicode("").tag(sync=True)
     subset_items = List([]).tag(sync=True)
+    subset_selected = Unicode("").tag(sync=True)
+    bg_subset_items = List(['Manual']).tag(sync=True)
+    bg_subset_selected = Unicode("Manual").tag(sync=True)
     background_value = Any(0).tag(sync=True)
     pixel_area = Any(0).tag(sync=True)
     counts_factor = Any(0).tag(sync=True)
@@ -62,11 +66,14 @@ class SimpleAperturePhotometry(TemplateMixin):
 
         self.subset_items = [lyr.label for lyr in self.app.data_collection.subset_groups
                              if lyr.label.startswith('Subset')]
+        self.bg_subset_items = ['Manual'] + self.subset_items
 
-    def vue_data_selected(self, event):
+    @observe('data_selected')
+    def _data_selected_changed(self, event={}):
+        data_selected = event.get('new', self.data_selected)
         try:
             self._selected_data = self.app.data_collection[
-                self.app.data_collection.labels.index(event)]
+                self.app.data_collection.labels.index(data_selected)]
             self.counts_factor = 0
             self.pixel_area = 0
             self.flux_scaling = 0
@@ -103,32 +110,61 @@ class SimpleAperturePhotometry(TemplateMixin):
             self.reset_results()
             self._selected_data = None
             self.hub.broadcast(SnackbarMessage(
-                f"Failed to extract {event}: {repr(e)}", color='error', sender=self))
+                f"Failed to extract {data_selected}: {repr(e)}", color='error', sender=self))
 
-    def vue_subset_selected(self, event):
+        # Auto-populate background, if applicable
+        self._bg_subset_selected_changed()
+
+    def _get_region_from_subset(self, subset):
+        for subset_grp in self.app.data_collection.subset_groups:
+            if subset_grp.label == subset:
+                for sbst in subset_grp.subsets:
+                    if sbst.data.label == self.data_selected:
+                        # TODO: https://github.com/glue-viz/glue-astronomy/issues/52
+                        return sbst.data.get_selection_definition(
+                                subset_id=subset, format='astropy-regions')
+        else:
+            raise ValueError(f'Subset "{subset}" not found')
+
+    @observe('subset_selected')
+    def _subset_selected_changed(self, event={}):
+        subset_selected = event.get('new', self.subset_selected)
         if self._selected_data is None:
             self.reset_results()
             return
 
-        subset = None
         try:
-            for subset_grp in self.app.data_collection.subset_groups:
-                if subset_grp.label == event:
-                    for sbst in subset_grp.subsets:
-                        if sbst.data.label == self._selected_data.label:
-                            subset = sbst
-                            break
-                    break
-
-            # TODO: https://github.com/glue-viz/glue-astronomy/issues/52
-            self._selected_subset = subset.data.get_selection_definition(
-                subset_id=event, format='astropy-regions')
-            self._selected_subset.meta['label'] = subset.label
+            self._selected_subset = self._get_region_from_subset(subset_selected)
+            self._selected_subset.meta['label'] = subset_selected
         except Exception as e:
             self._selected_subset = None
             self.reset_results()
             self.hub.broadcast(SnackbarMessage(
-                f"Failed to extract {event}: {repr(e)}", color='error', sender=self))
+                f"Failed to extract {subset_selected}: {repr(e)}", color='error', sender=self))
+
+    @observe('bg_subset_selected')
+    def _bg_subset_selected_changed(self, event={}):
+        bg_subset_selected = event.get('new', self.bg_subset_selected)
+        if bg_subset_selected == 'Manual':
+            # we'll later access the user's self.background_value directly
+            return
+
+        try:
+            # Basically same way image stats are calculated in vue_do_aper_phot()
+            # except here we only care about one stat for the background.
+            data = self._selected_data
+            reg = self._get_region_from_subset(bg_subset_selected)
+            comp = data.get_component(data.main_components[0])
+            aper_mask_stat = reg.to_mask(mode='center')
+            img_stat = aper_mask_stat.get_values(comp.data, mask=None)
+
+            # photutils/background/_utils.py --> nanmedian()
+            self.background_value = np.nanmedian(img_stat)  # Naturally in data unit
+
+        except Exception as e:
+            self.background_value = 0
+            self.hub.broadcast(SnackbarMessage(
+                f"Failed to extract {bg_subset_selected}: {repr(e)}", color='error', sender=self))
 
     def vue_do_aper_phot(self, *args, **kwargs):
         if self._selected_data is None or self._selected_subset is None: